Reworked logic - show notification only if is timer running iin the background
