+++ app/build.gradle
@@ -76,4 +76,5 @@ dependencies {
     implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-core:1.3.3'
     implementation 'androidx.lifecycle:lifecycle-runtime-ktx:2.2.0'
     implementation 'org.greenrobot:eventbus:3.2.0'
+    implementation 'androidx.lifecycle:lifecycle-extensions:2.2.0'
 }
