+++ app/src/main/kotlin/com/simplemobiletools/clock/receivers/HideTimerReceiver.kt
@@ -4,9 +4,12 @@ import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
 import com.simplemobiletools.clock.extensions.hideTimerNotification
+import com.simplemobiletools.clock.services.TimerState
+import org.greenrobot.eventbus.EventBus
 
 class HideTimerReceiver : BroadcastReceiver() {
     override fun onReceive(context: Context, intent: Intent) {
         context.hideTimerNotification()
+        EventBus.getDefault().post(TimerState.Idle)
     }
 }
